pyproject.toml
[project]
name = "pss-fleet-data-client"
description = "A client for the Pixel Starships Fleet Data API."
version = "0.4.0"
requires-python = ">= 3.11"
license = {file = "LICENSE"}
authors = [
    {name = "The worst.", email = "theworstpss@gmail.com"}
]
readme = "README.md"
keywords = [
    "Pixel Starships",
    "PSS",
    "Fleet Data",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: AsyncIO",
    "Framework :: Pydantic",
    "Framework :: Pydantic :: 2",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27.0",
    "pssapi>=0.5.1",
    "pydantic>=2.8.2",
    "python-dateutil>=2.9.0.post0",
]

[project.urls]
Repository = "https://github.com/PSS-Tools-Development/pss-fleet-data-client"
Issues = "https://github.com/PSS-Tools-Development/pss-fleet-data-client/issues"
Changelog = "https://github.com/PSS-Tools-Development/pss-fleet-data-client/releases"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/pss_fleet_data/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = [
    "src/pss_fleet_data",
]

[tool.hatch.build.targets.sdist]
include = [
  "src/pss_fleet_data/*",
]
exclude = [
  "src/pss_fleet_data/__pycache__",
  "src/pss_fleet_data/pss_fleet_data_client.egg-info",
]

[tool.rye]
managed = true
dev-dependencies = [
    "autoflake>=2.3.1",
    "black>=24.4.2",
    "flake8>=7.1.0",
    "flake8-black>=0.3.6",
    "flake8-bugbear>=24.4.26",
    "isort>=5.13.2",
    "pytest>=7.2.1",
    "pytest-asyncio>=0.23.7",
    "pytest-env>=1.1.3",
    "pytest_httpx>=0.30.0",
    "pytest-vcr>=1.0.2",
    "vulture>=2.11",
    "pre-commit>=3.8.0",
    "pytest-cov>=5.0.0",
    "coverage>=7.6.0",
]

[tool.autoflake]
in-place = true
remove-all-unused-imports = true
ignore-init-module-imports = true
recursive = true

[tool.black]
line-length = 150
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
src_paths = ["./src", "./tests"]
profile = "black"
line_length = 150
lines_after_imports = 2

[tool.vulture]
min_confidence = 100
paths = ["./src/pss_fleet_data"]
sort_by_size = true
